src/requirements.txt
werkzeug==0.16.0
flask
gunicorn
tornado
numpy
pandas
pillow
pyyaml
opencv-python
requests
matplotlib
pycryptodome==3.9.9
